Move side scroll functionality to MediaSideScroll.kt
